refactor bot structure
